Moved dashboard to home page
